// packages/nx-plugin/src/agentcore-gateway/react-connection/schema.d.ts
/**
 * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 * SPDX-License-Identifier: Apache-2.0
 */
import type { ComponentMetadata } from '../../utils/nx.js';

export interface AgentcoreGatewayReactConnectionGeneratorSchema {
  sourceProject: string;
  targetProject: string;
  sourceComponent?: ComponentMetadata;
  targetComponent?: ComponentMetadata;
  preferInstallDependencies?: boolean;
}